refactor: Add Project module and update module configuration
This commit adds a new module called Project and updates the module configuration files. The `global.php` file in the `config/autoload` directory is modified to update the `db` key with a new adapter configuration. The `modules.config.php` file in the `config` directory is also updated to include the `Project` module.

Recent user commits:
- Merge pull request #2 from LucianTavares/1.7.x: refactor: Adiciona instalação do sqlite3 no Dockerfile.
- refactor: Adiciona instalação do sqlite3 no Dockerfile.
- docs: Improvements in README.md
- feat: Improvements in docker entrypoint
- fix: Execute in Docker using user non-root

Recent repository commits:
- Merge pull request #37 from weierophinney/feature/lcs-2.3: Adopt laminas-coding-standard 2.3 rules
- fix: bump laminas-test minimum to 3.5
- fix: do not test on PHP 7.3 with locked dependencies
- qa: apply laminas-coding-standard rules
- qa: update to laminas-coding-standard 2.3